+++ users/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v2.0.4] users
 ### Releases
 - [NPM v2.0.4] https://www.npmjs.com/package/@wingspan/users/v/2.0.4 - users
+
+## 2023-12-18 01:31:02
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.126.4 (2.214.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] users
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@wingspan/users/v/2.1.0 - users